requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
joblib
numpy
pandas
scipy
statsmodels
scikit-learn==1.5.2
xgboost
tensorflow
keras